+++ src/commands/copy.rs
@@ -2,40 +2,55 @@
 
 use crate::{
     commands::{get_repository, init::init_password, open_repository, open_repository_indexed},
-    config::AllRepositoryOptions,
     helpers::table_with_titles,
-    status_err, Application, RUSTIC_APP,
+    status_err, Application, RusticConfig, RUSTIC_APP,
 };
-use abscissa_core::{Command, Runnable, Shutdown};
+use abscissa_core::{config::Override, Command, FrameworkError, Runnable, Shutdown};
 use anyhow::{bail, Result};
-use log::{error, info};
-use merge::Merge;
+use conflate::Merge;
+use log::{error, info, log, Level};
 use serde::{Deserialize, Serialize};
 
 use rustic_core::{CopySnapshot, Id, KeyOptions};
 
 /// `copy` subcommand
-#[derive(clap::Parser, Command, Debug)]
-pub(crate) struct CopyCmd {
+#[derive(clap::Parser, Command, Default, Clone, Debug, Serialize, Deserialize, Merge)]
+pub struct CopyCmd {
     /// Snapshots to copy. If none is given, use filter options to filter from all snapshots.
     #[clap(value_name = "ID")]
+    #[serde(skip)]
+    #[merge(skip)]
     ids: Vec<String>,
 
     /// Initialize non-existing target repositories
     #[clap(long)]
+    #[serde(skip)]
+    #[merge(skip)]
     init: bool,
+
+    /// Target repository (can be specified multiple times)
+    #[clap(long = "target", value_name = "TARGET")]
+    #[merge(strategy=conflate::vec::overwrite_empty)]
+    targets: Vec<String>,
 
     /// Key options (when using --init)
     #[clap(flatten, next_help_heading = "Key options (when using --init)")]
+    #[serde(skip)]
+    #[merge(skip)]
     key_opts: KeyOptions,
 }
 
-/// Target repository options
-#[derive(Default, Clone, Debug, Serialize, Deserialize, Merge)]
-pub struct Targets {
-    /// Target repositories
-    #[merge(strategy = merge::vec::overwrite_empty)]
-    targets: Vec<AllRepositoryOptions>,
+impl Override<RusticConfig> for CopyCmd {
+    // Process the given command line options, overriding settings from
+    // a configuration file using explicit flags taken from command-line
+    // arguments.
+    fn override_config(&self, mut config: RusticConfig) -> Result<RusticConfig, FrameworkError> {
+        let mut self_config = self.clone();
+        // merge "copy" section from config file, if given
+        self_config.merge(config.copy);
+        config.copy = self_config;
+        Ok(config)
+    }
 }
 
 impl Runnable for CopyCmd {
@@ -52,8 +67,7 @@ impl CopyCmd {
         let config = RUSTIC_APP.config();
 
         if config.copy.targets.is_empty() {
-            status_err!("no [[copy.targets]] section in config file found!");
-            RUSTIC_APP.shutdown(Shutdown::Crash);
+            bail!("No target given. Please specify at least 1 target either in the profile or using --target!");
         }
 
         let repo = open_repository_indexed(&config.repository)?;
@@ -66,7 +80,16 @@ impl CopyCmd {
         snapshots.sort_unstable();
 
         let poly = repo.config().poly()?;
-        for target_opt in &config.copy.targets {
+        for target in &config.copy.targets {
+            let mut merge_logs = Vec::new();
+            let mut target_config = RusticConfig::default();
+            target_config.merge_profile(target, &mut merge_logs, Level::Error)?;
+            // display logs from merging
+            for (level, merge_log) in merge_logs {
+                log!(level, "{}", merge_log);
+            }
+            let target_opt = &target_config.repository;
+
             let repo_dest = get_repository(target_opt)?;
 
             info!("copying to target {}...", repo_dest.name);
@@ -79,7 +102,7 @@ impl CopyCmd {
                     continue;
                 }
                 let mut config_dest = repo.config().clone();
-                config_dest.id = Id::random();
+                config_dest.id = Id::random().into();
                 let pass = init_password(&repo_dest)?;
                 repo_dest.init_with_config(&pass, &self.key_opts, config_dest)?
             } else {