+++ src/commands/forget.rs
@@ -10,7 +10,7 @@ use abscissa_core::{Command, FrameworkError, Runnable};
 use anyhow::Result;
 
 use chrono::Local;
-use merge::Merge;
+use conflate::Merge;
 use serde::{Deserialize, Serialize};
 use serde_with::{serde_as, DisplayFromStr};
 
@@ -70,11 +70,12 @@ pub struct ForgetOptions {
     /// Group snapshots by any combination of host,label,paths,tags (default: "host,label,paths")
     #[clap(long, short = 'g', value_name = "CRITERION")]
     #[serde_as(as = "Option<DisplayFromStr>")]
+    #[merge(strategy=conflate::option::overwrite_none)]
     group_by: Option<SnapshotGroupCriterion>,
 
     /// Also prune the repository
     #[clap(long)]
-    #[merge(strategy = merge::bool::overwrite_false)]
+    #[merge(strategy=conflate::bool::overwrite_false)]
     prune: bool,
 
     /// Snapshot filter options
@@ -98,6 +99,9 @@ impl Runnable for ForgetCmd {
 }
 
 impl ForgetCmd {
+    /// be careful about self vs `RUSTIC_APP.config()` usage
+    /// only the `RUSTIC_APP.config()` involves the TOML and ENV merged configurations
+    /// see <https://github.com/rustic-rs/rustic/issues/1242>
     fn inner_run(&self) -> Result<()> {
         let config = RUSTIC_APP.config();
         let repo = open_repository(&config.repository)?;
@@ -155,7 +159,7 @@ impl ForgetCmd {
             (_, _, true) => {}
         }
 
-        if self.config.prune {
+        if config.forget.prune {
             let mut prune_opts = self.prune_opts.clone();
             prune_opts.opts.ignore_snaps = forget_snaps;
             prune_opts.run();