+++ src/commands/webdav.rs
@@ -8,8 +8,8 @@ use std::{net::ToSocketAddrs, str::FromStr};
 use crate::{commands::open_repository_indexed, status_err, Application, RusticConfig, RUSTIC_APP};
 use abscissa_core::{config::Override, Command, FrameworkError, Runnable, Shutdown};
 use anyhow::{anyhow, Result};
+use conflate::Merge;
 use dav_server::{warp::dav_handler, DavHandler};
-use merge::Merge;
 use serde::{Deserialize, Serialize};
 
 use rustic_core::vfs::{FilePolicy, IdenticalSnapshot, Latest, Vfs};
@@ -19,27 +19,32 @@ use rustic_core::vfs::{FilePolicy, IdenticalSnapshot, Latest, Vfs};
 pub struct WebDavCmd {
     /// Address to bind the webdav server to. [default: "localhost:8000"]
     #[clap(long, value_name = "ADDRESS")]
+    #[merge(strategy=conflate::option::overwrite_none)]
     address: Option<String>,
 
     /// The path template to use for snapshots. {id}, {id_long}, {time}, {username}, {hostname}, {label}, {tags}, {backup_start}, {backup_end} are replaced. [default: "[{hostname}]/[{label}]/{time}"]
     #[clap(long)]
+    #[merge(strategy=conflate::option::overwrite_none)]
     path_template: Option<String>,
 
     /// The time template to use to display times in the path template. See https://docs.rs/chrono/latest/chrono/format/strftime/index.html for format options. [default: "%Y-%m-%d_%H-%M-%S"]
     #[clap(long)]
+    #[merge(strategy=conflate::option::overwrite_none)]
     time_template: Option<String>,
 
     /// Use symlinks. This may not be supported by all WebDAV clients
     #[clap(long)]
-    #[merge(strategy = merge::bool::overwrite_false)]
+    #[merge(strategy=conflate::bool::overwrite_false)]
     symlinks: bool,
 
     /// How to handle access to files. [default: "forbidden" for hot/cold repositories, else "read"]
     #[clap(long)]
+    #[merge(strategy=conflate::option::overwrite_none)]
     file_access: Option<String>,
 
     /// Specify directly which snapshot/path to serve
     #[clap(value_name = "SNAPSHOT[:PATH]")]
+    #[merge(strategy=conflate::option::overwrite_none)]
     snapshot_path: Option<String>,
 }
 
@@ -66,27 +71,32 @@ impl Runnable for WebDavCmd {
 }
 
 impl WebDavCmd {
+    /// be careful about self VS RUSTIC_APP.config() usage
+    /// only the RUSTIC_APP.config() involves the TOML and ENV merged configurations
+    /// see https://github.com/rustic-rs/rustic/issues/1242
     fn inner_run(&self) -> Result<()> {
         let config = RUSTIC_APP.config();
         let repo = open_repository_indexed(&config.repository)?;
 
-        let path_template = self
+        let path_template = config
+            .webdav
             .path_template
             .clone()
             .unwrap_or_else(|| "[{hostname}]/[{label}]/{time}".to_string());
-        let time_template = self
+        let time_template = config
+            .webdav
             .time_template
             .clone()
             .unwrap_or_else(|| "%Y-%m-%d_%H-%M-%S".to_string());
 
         let sn_filter = |sn: &_| config.snapshot_filter.matches(sn);
 
-        let vfs = if let Some(snap) = &self.snapshot_path {
+        let vfs = if let Some(snap) = &config.webdav.snapshot_path {
             let node = repo.node_from_snapshot_path(snap, sn_filter)?;
             Vfs::from_dir_node(&node)
         } else {
             let snapshots = repo.get_matching_snapshots(sn_filter)?;
-            let (latest, identical) = if self.symlinks {
+            let (latest, identical) = if config.webdav.symlinks {
                 (Latest::AsLink, IdenticalSnapshot::AsLink)
             } else {
                 (Latest::AsDir, IdenticalSnapshot::AsDir)
@@ -94,7 +104,8 @@ impl WebDavCmd {
             Vfs::from_snapshots(snapshots, &path_template, &time_template, latest, identical)?
         };
 
-        let addr = self
+        let addr = config
+            .webdav
             .address
             .clone()
             .unwrap_or_else(|| "localhost:8000".to_string())
@@ -102,7 +113,7 @@ impl WebDavCmd {
             .next()
             .ok_or_else(|| anyhow!("no address given"))?;
 
-        let file_access = self.file_access.as_ref().map_or_else(
+        let file_access = config.webdav.file_access.as_ref().map_or_else(
             || {
                 if repo.config().is_hot == Some(true) {
                     Ok(FilePolicy::Forbidden)