+++ xtask/src/lib.rs
@@ -1,4 +1,0 @@
-pub mod helpers;
-pub mod tasks;
-
-pub use tasks::install_deps::InstallationKind;