+++ CHANGELOG.md
@@ -2,6 +2,109 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.9.0](https://github.com/rustic-rs/rustic/compare/v0.8.1...v0.9.0) - 2024-09-29
+
+### Bug Fixes
+
+- [**breaking**] use multiple options only as array in config profile
+  ([#1240](https://github.com/rustic-rs/rustic/pull/1240))
+- Allow snapshots to be modified and marked to forget
+  ([#1253](https://github.com/rustic-rs/rustic/pull/1253))
+- make ls and find show the year of mtime date
+  ([#1249](https://github.com/rustic-rs/rustic/pull/1249))
+- ls: Remove printing trailing space
+  ([#1247](https://github.com/rustic-rs/rustic/pull/1247))
+- webdav/forget: correctly use application config
+  ([#1241](https://github.com/rustic-rs/rustic/pull/1241))
+
+### Features
+
+- [**breaking**] copy: Use config profile as target
+  ([#1131](https://github.com/rustic-rs/rustic/pull/1131))
+- backup: Add option `stdin-command` 
+  ([rustic_core](https://github.com/rustic-rs/rustic_core/releases/tag/rustic_core-v0.3.0))
+- Add list indexpacks and list indexcontent commands
+  ([#1254](https://github.com/rustic-rs/rustic/pull/1254))
+- Add option `--only-identical` for `diff` to allow for bitrot check
+  ([#1250](https://github.com/rustic-rs/rustic/pull/1250))
+- ls: Add option --json ([#1251](https://github.com/rustic-rs/rustic/pull/1251))
+- backup: Add option `--long`
+  ([#1159](https://github.com/rustic-rs/rustic/pull/1159))
+
+### Documentation
+
+- update installation instructions in readme to use `--locked` flag for install
+  from crates.io
+- update RepositoryErrorKind rustdoc following rustic_core change
+  ([#1237](https://github.com/rustic-rs/rustic/pull/1237))
+
+### Other
+
+- Remove self-update from default crate features
+  ([#1139](https://github.com/rustic-rs/rustic/pull/1139))
+- Reduce memory usage of restore
+  ([#1069](https://github.com/rustic-rs/rustic/pull/1069))
+- *(deps)* update rust crate libc to v0.2.159
+  ([#1257](https://github.com/rustic-rs/rustic/pull/1257))
+- *(deps)* lock file maintenance
+  ([#1269](https://github.com/rustic-rs/rustic/pull/1269))
+- *(deps)* update rust crate rstest to 0.23
+  ([#1267](https://github.com/rustic-rs/rustic/pull/1267))
+- *(deps)* update rust crate tempfile to v3.13.0
+  ([#1266](https://github.com/rustic-rs/rustic/pull/1266))
+- *(deps)* update marcoieni/release-plz-action digest to 8b0f89a
+  ([#1265](https://github.com/rustic-rs/rustic/pull/1265))
+- *(deps)* update embarkstudios/cargo-deny-action action to v2
+  ([#1259](https://github.com/rustic-rs/rustic/pull/1259))
+- *(deps)* update rustsec/audit-check action to v2
+  ([#1260](https://github.com/rustic-rs/rustic/pull/1260))
+- *(deps)* update softprops/action-gh-release action to v2
+  ([#1258](https://github.com/rustic-rs/rustic/pull/1258))
+- *(deps)* update embarkstudios/cargo-deny-action digest to 3f4a782
+  ([#1228](https://github.com/rustic-rs/rustic/pull/1228))
+
+## [0.8.1] - 2024-09-08
+
+### Bug Fixes
+
+- Allow to compile without tui feature
+  ([#1208](https://github.com/rustic-rs/rustic/issues/1208))
+- Use cargo --locked in CI pipeline
+  ([#1207](https://github.com/rustic-rs/rustic/issues/1207))
+- Return exitcode ([#1220](https://github.com/rustic-rs/rustic/issues/1220))
+- "Incorrect Password" error is now only shown if password is really incorrect.
+  ([rustic_core](https://github.com/rustic-rs/rustic_core/releases/tag/rustic_core-v0.3.1))
+- Group by now works as expected
+  ([rustic_core](https://github.com/rustic-rs/rustic_core/releases/tag/rustic_core-v0.3.1))
+- A bug in `keep-tags` and `filter-tags` has been fixed.
+  ([rustic_core](https://github.com/rustic-rs/rustic_core/releases/tag/rustic_core-v0.3.1))
+- Building OpenBSD platform target is now possible again
+  ([rustic_core](https://github.com/rustic-rs/rustic_core/releases/tag/rustic_core-v0.3.1))
+
+### Documentation
+
+- Update config profile readme
+  ([#1221](https://github.com/rustic-rs/rustic/issues/1221))
+
+### Features
+
+- Add autocompletion hints
+  ([#1225](https://github.com/rustic-rs/rustic/issues/1225))
+- Allow to modify filters
+  ([#1210](https://github.com/rustic-rs/rustic/issues/1210))
+- Allow to view text files
+  ([#1216](https://github.com/rustic-rs/rustic/issues/1216))
+
+### Generated
+
+- Updated Completions fixtures
+
+### Miscellaneous Tasks
+
+- Bump quinn-proto from 0.11.6 to 0.11.8
+  ([#1223](https://github.com/rustic-rs/rustic/issues/1223))
+- Dependency updates ([#1227](https://github.com/rustic-rs/rustic/issues/1227))
+
 ## [0.8.0] - 2024-08-21
 
 ### Bug Fixes